+++ lib/code/parser/statement.rb
@@ -1,7 +1,7 @@
 class Code
   class Parser
     class Statement < Parslet::Parser
-      rule(:statement) { ::Code::Parser::While.new }
+      rule(:statement) { ::Code::Parser::Equality.new }
       root(:statement)
     end
   end