+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds `bin/code` with same options as `bin/template`
 - Prevent loose syntax like `{ a: }`, `[1,,,]` and `()`
 - Change precedence of defined? (to allow `defined?(name) ? name : nothing`)
+- Updates parsers to allow `while false end == nothing`
 
 ## 0.2.4 / 2022-08-02